requirements.txt
pytorch-lightning==1.3.2
transformers==4.6.0
pyarrow==2.0.0
timm==0.4.12
Pillow==8.1.0
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
einops==0.3.0
sacred==0.8.2
pandas==1.1.5
ftfy==6.0.3
torchmetrics==0.7.2
scikit-learn==1.0.2
wandb